Rename files for consistent naming.
