+++ pyBeam/src/element.cpp
@@ -25,7 +25,7 @@
  */
 
 
-#include "../include/FiniteElement.h"
+#include "../include/element.h"
 #include <iostream>
 
 CElement::CElement(void){
